requirements.txt
python-binance==1.0.15
sqlalchemy==1.4.41
psycopg2==2.9.4
configparser==5.0.2
asyncio==3.4.3
pyyaml==6.0
requests==2.28.1